+++ CHANGELOG.md
@@ -17,6 +17,10 @@
 
 - added `isNotNull`
 
+### Array -> (String -> Boolean)
+
+- added `containsOneOf`
+
 ### Object -> Any
 
 - added `getId`